refactor: 混雑状況のデータ型を int から int8 に変更
- Booth ドメインモデルの congestionStatus フィールドおよび関連メソッドの型を int8 に更新
- リポジトリ、ユースケース、テストコードにおける関連する引数や戻り値の型を修正

+++ internal/domain/booth.go
@@ -14,7 +14,7 @@ type Booth struct {
 	Name             string  // ブース名
 	Organizer        string  // ブースの主催者(ex. 1-1, 陸上部...)
 	Detail           string  // ブースの説明
-	congestionStatus int     // 0: 空き 1: 少し混雑している 2: かなり混雑している]
+	congestionStatus int8     // 0: 空き 1: 少し混雑している 2: かなり混雑している]
 	X                float32 // X座標
 	Y                float32 // Y座標
 	Z                float32 // Z座標
@@ -48,15 +48,15 @@ type BoothRepository interface {
 	Create(ctx context.Context, booth *Booth) error
 	GetByID(ctx context.Context, id string) (*Booth, error)
 	Update(ctx context.Context, booth *Booth) error
-	UpdateCongestion(ctx context.Context, id string, congestionLevel int) error 
+	UpdateCongestion(ctx context.Context, id string, congestionLevel int8) error 
 	GetAllBooths(ctx context.Context) ([]*Booth, error)
 }
 
-func (b *Booth) CongestionStatus() int {
+func (b *Booth) CongestionStatus() int8 {
 	return b.congestionStatus
 }
 
-func (b *Booth) SetCongestionStatus(congestionLevel int) error {
+func (b *Booth) SetCongestionStatus(congestionLevel int8) error {
 	if err := ValidateCongestionLevel(congestionLevel); err != nil {
 		return err
 	}
@@ -64,7 +64,7 @@ func (b *Booth) SetCongestionStatus(congestionLevel int) error {
 	return nil
 }
 
-func ValidateCongestionLevel(congestionLevel int) error {
+func ValidateCongestionLevel(congestionLevel int8) error {
 	if 0 > congestionLevel || congestionLevel > 2 {
 		return errors.New("Congestion level must be between 0 and 2")
 	}
